# PERF: _return_parsed_timezone_results