--- a/pandas/core/tools/datetimes.py
+++ b/pandas/core/tools/datetimes.py
@@ -313,19 +313,17 @@ def _return_parsed_timezone_results(
     -------
     tz_result : Index-like of parsed dates with timezone
     """
-    tz_results = np.array(
-        [Timestamp(res).tz_localize(zone) for res, zone in zip(result, timezones)]
-    )
-    if utc:
-        # Convert to the same tz
-        tz_results = np.array(
-            [
-                tz_result.tz_convert("utc")
-                if tz_result.tzinfo is not None
-                else tz_result.tz_localize("utc")
-                for tz_result in tz_results
-            ]
-        )
+    dta = DatetimeArray(result)
+    tz_results = np.empty(len(dta), dtype=object)
+    for zone in unique(timezones):
+        mask = timezones == zone
+        dta = DatetimeArray(result[mask]).tz_localize(zone)
+        if utc:
+            if dta.tzinfo is None:
+                dta = dta.tz_localize("utc")
+            else:
+                dta = dta.tz_convert("utc")
+        tz_results[mask] = dta
 
     return Index(tz_results, name=name)