- [x] closes #50107 (Replace xxxx with the GitHub issue number)
- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.0.0.rst` file if fixing a bug or adding a new feature.

@MarcoGorelli - thanks for the ping. I think this gets most of the bottleneck you highlighted:

```
import pandas as pd

dates = pd.date_range('1900', '2000').tz_localize('+01:00').strftime('%Y-%d-%m %H:%M:%S%z').tolist()
dates.append('2020-01-01 00:00:00+02:00')

%timeit pd.to_datetime(dates, format='%Y-%d-%m %H:%M:%S%z')

# 529 ms ± 24.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   <- main
# 174 ms ± 977 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  <- PR
```